Display for wlp3s0 works now need to stagger
I need to focus on staggering for performance
Currently I have 3 sensors and they all trigger at the
same time and return data roughly at the same time.
The more sensors the more of a spike it is to process the
data for the sensors and update the graphs.
It would be ideal to have a way to intelligently trigger
the sensors so they don't all fire at the same time.
Possibly instead of using the tick timer call for updates
another server or supervisor could coordinate when to
call the sensors for triggering updates.
This could be a supervisor/genserver combo (pool style).
This way sensors will update at different times keeping
things from spiking and surging as more sensors are added
to the system and more graphs are added to the dashboard.

@tapickell